+++ .gitignore
@@ -6,3 +6,4 @@ proxyctl
 ext.yml
 tuf-repo-cdn.sigstore.dev.json
 router
+logs/
