+++ toolruntime/chat_stream.go
@@ -16,163 +16,6 @@ import (
 	"github.com/tinfoilsh/confidential-model-router/manager"
 )
 
-// runChatStreaming is the streaming entry point for the Chat Completions
-// route. It drives the same tool loop as runChatLoop but consumes the
-// upstream model's stream incrementally so that assistant content deltas
-// and citation annotations are forwarded to the client live rather than
-// synthesized at the end of the turn. When the client opted into the
-// router-owned `X-Tinfoil-Events: web_search` marker stream, router-owned
-// web_search tool-call progress is surfaced inline via
-// `<tinfoil-event>...</tinfoil-event>` markers carried inside the normal
-// `delta.content` text. Clients that did not opt in see a pristine
-// spec-conformant chat.completion.chunk stream with no markers.
-//
-// The function assumes isStream(body) is already true. Non-streaming callers
-// continue to go through runChatLoop so existing behavior is preserved for
-// SDKs that do not want SSE back.
-func runChatStreaming(
-	ctx context.Context,
-	w http.ResponseWriter,
-	r *http.Request,
-	em *manager.EnclaveManager,
-	registry *sessionRegistry,
-	body map[string]any,
-	modelName string,
-	requestHeaders http.Header,
-	prompt *mcp.GetPromptResult,
-) error {
-	flusher, ok := w.(http.Flusher)
-	if !ok {
-		return fmt.Errorf("streaming not supported")
-	}
-
-	tid := debugTraceID()
-	searchOpts := parseChatWebSearchOptions(body)
-	tools := registry.allTools()
-	ownedTools := registry.ownedTools()
-	toolSchemas := schemaLookup(tools)
-	reqBody := cloneJSONMap(body)
-	delete(reqBody, "web_search_options")
-	delete(reqBody, "filters")
-	delete(reqBody, "pii_check_options")
-	delete(reqBody, "prompt_injection_check_options")
-	stripRouterOwnedIncludes(reqBody)
-	reqBody["stream"] = true
-	reqBody["stream_options"] = map[string]any{"include_usage": true}
-	applyParallelToolCallsPolicy(reqBody)
-	reqBody["tools"] = append(existingTools(reqBody["tools"]), chatTools(tools)...)
-	reqBody["messages"] = prependChatPrompt(prompt, reqBody["messages"])
-
-	usageMetricsRequested := r.Header.Get(manager.UsageMetricsRequestHeader) == "true"
-	clientRequestedUsage := r.Header.Get("X-Tinfoil-Client-Requested-Usage") == "true"
-	eventsEnabled := tinfoilEventsEnabled(r.Header)
-
-	if debugEnabled {
-		ownedNames := make([]string, 0, len(ownedTools))
-		for n := range ownedTools {
-			ownedNames = append(ownedNames, n)
-		}
-		debugLogf("toolruntime:%s chatstream.start model=%s search_ctx=%q profiles=%v ownedTools=%v schemas=%d",
-			tid, modelName, searchOpts.searchContextSize, registry.profileNames(), ownedNames, len(toolSchemas))
-	}
-
-	streamer := &chatStreamer{
-		streamBase: streamBase{
-			w:                     w,
-			flusher:               flusher,
-			usageMetricsRequested: usageMetricsRequested,
-			citations:             &citationState{nextIndex: 1},
-			usageTotals:           &usageAccumulator{},
-		},
-		clientRequestedUsage: clientRequestedUsage,
-		eventsEnabled:        eventsEnabled,
-		emitter:              nil,
-	}
-	streamer.emitter = newCitationEmitter(streamer.citations)
-
-	for i := 0; i < maxToolIterations; i++ {
-		if msgs, ok := reqBody["messages"].([]any); ok {
-			if n := sanitizeAssistantToolCallsInMessages(msgs); n > 0 {
-				debugLogf("toolruntime:%s chatstream.iter=%d history_sanitized count=%d", tid, i, n)
-			}
-		}
-		if debugEnabled {
-			msgs, _ := reqBody["messages"].([]any)
-			debugLogf("toolruntime:%s chatstream.iter=%d upstream.request messages_n=%d history=%s", tid, i, len(msgs), debugMessagesSummary(msgs, 200))
-		}
-		iterStart := time.Now()
-		result, err := streamer.runIteration(ctx, em, modelName, reqBody, requestHeaders)
-		if err != nil {
-			debugLogf("toolruntime:%s chatstream.iter=%d upstream.error elapsed=%s err=%v", tid, i, time.Since(iterStart), err)
-			return streamer.terminateWithError(r, em, modelName, err)
-		}
-		// If the client has disconnected, stop before spending another
-		// round of MCP tool calls and upstream tokens on work nobody is
-		// listening for.
-		if streamer.writeErr != nil {
-			debugLogf("toolruntime:%s chatstream.client_disconnected at iter=%d", tid, i)
-			streamer.emitBillingEvent(r, em, modelName, streamer.finalUsage())
-			return nil
-		}
-
-		routerToolCalls, clientToolCalls := splitToolCalls(ownedTools, result.toolCalls)
-		if debugEnabled {
-			toolNames := make([]string, 0, len(result.toolCalls))
-			for _, c := range result.toolCalls {
-				toolNames = append(toolNames, c.name)
-			}
-			debugLogf("toolruntime:%s chatstream.iter=%d upstream.response elapsed=%s finish=%q total_tool_calls=%d router_tool_calls=%d client_tool_calls=%d tool_names=%v content=%q",
-				tid, i, time.Since(iterStart), result.finishReason, len(result.toolCalls), len(routerToolCalls), len(clientToolCalls), toolNames, debugPreview(result.content, 240))
-		}
-		if len(routerToolCalls) == 0 {
-			debugLogf("toolruntime:%s chatstream.done iterations=%d (no router tool calls)", tid, i+1)
-			return streamer.finalize(r, em, modelName, result, clientToolCalls)
-		}
-
-		mixedTurn := len(clientToolCalls) > 0
-		tracePhase := fmt.Sprintf("chatstream.iter=%d", i)
-		var messages []any
-		if !mixedTurn {
-			messages, _ = reqBody["messages"].([]any)
-			messages = append(messages, result.assistantMessage())
-		}
-		for _, call := range routerToolCalls {
-			output := resolveStreamingRouterToolCall(
-				ctx, call, searchOpts, toolSchemas, streamer.citations,
-				func(ctx context.Context, call toolCall) (string, error) {
-					return streamer.executeTool(ctx, registry, call)
-				},
-				tracePhase, tid,
-			)
-			if !mixedTurn {
-				messages = append(messages, map[string]any{
-					"role":         "tool",
-					"tool_call_id": call.id,
-					"content":      output,
-				})
-			}
-		}
-		// Mixed turn: see the mixed-turn contract on runToolLoop.
-		if mixedTurn {
-			debugLogf("toolruntime:%s chatstream.mixed_turn iterations=%d (router+client calls in one turn; finalizing)", tid, i+1)
-			return streamer.finalize(r, em, modelName, result, clientToolCalls)
-		}
-		reqBody["messages"] = messages
-	}
-	debugLogf("toolruntime:%s chatstream.exhausted iterations=%d forcing final answer", tid, maxToolIterations)
-
-	// Exhausted the tool budget; force a final-answer turn with tools removed
-	// and stream it to the client directly.
-	finalBody := forcedFinalChatRequest(reqBody)
-	finalBody["stream"] = true
-	finalBody["stream_options"] = map[string]any{"include_usage": true}
-	result, err := streamer.runIteration(ctx, em, modelName, finalBody, requestHeaders)
-	if err != nil {
-		return streamer.terminateWithError(r, em, modelName, err)
-	}
-	return streamer.finalize(r, em, modelName, result, nil)
-}
-
 // chatStreamer owns one logical chat.completion.chunk stream across
 // every tool iteration, keeping id/created/model pinned to the first
 // upstream chunk's values.
@@ -180,10 +23,10 @@ type chatStreamer struct {
 	streamBase
 
 	clientRequestedUsage bool
-	// eventsEnabled is true when the caller opted into the
-	// `X-Tinfoil-Events: web_search` marker stream.
-	eventsEnabled bool
-	emitter       *citationEmitter
+	// eventFlags tracks which tinfoil-event marker families the
+	// caller opted into via the `X-Tinfoil-Events` header.
+	eventFlags tinfoilEventFlags
+	emitter    *citationEmitter
 
 	// roleEmitted defers the initial assistant role-delta chunk until
 	// the first upstream frame, so it can carry upstream's identity.
@@ -205,11 +48,32 @@ type chatStreamer struct {
 // only client-owned tool calls remain).
 type chatIterationResult struct {
 	content      string
+	reasoning    string
 	toolCalls    []toolCall
 	rawToolCalls []any
 	finishReason string
 	usage        map[string]any
 }
+
+// chatToolCallBuilder assembles OpenAI streaming tool_call deltas into the
+// final tool_call objects. OpenAI emits each function's name and id on
+// the first delta for a given index, and the arguments JSON in
+// incremental string fragments.
+type chatToolCallBuilder struct {
+	entries []*chatToolCallEntry
+}
+
+type chatToolCallEntry struct {
+	index        int
+	id           string
+	toolType     string
+	functionName string
+	arguments    []byte
+}
+
+// ---------------------------------------------------------------------------
+// chatIterationResult methods
+// ---------------------------------------------------------------------------
 
 // assistantMessage builds the OpenAI-shaped assistant message that the
 // router appends to the messages array between iterations so the next
@@ -226,6 +90,10 @@ func (r *chatIterationResult) assistantMessage() map[string]any {
 	}
 	return message
 }
+
+// ---------------------------------------------------------------------------
+// chatStreamer methods
+// ---------------------------------------------------------------------------
 
 // runIteration posts one upstream streaming call and drives the SSE pump
 // until the upstream signals done. Content deltas flow through the citation
@@ -321,6 +189,11 @@ func (s *chatStreamer) pumpUpstream(reader *sseReader) (chatIterationResult, err
 		// citations so they bypass the citation emitter. They must still
 		// flow through to the client as-is so reasoning models keep
 		// driving the live thinking UI when web search is enabled.
+		if r := stringValue(delta["reasoning_content"]); r != "" {
+			result.reasoning += r
+		} else if r := stringValue(delta["reasoning"]); r != "" {
+			result.reasoning += r
+		}
 		s.emitReasoningDelta(delta)
 		if rawCalls, ok := delta["tool_calls"].([]any); ok {
 			builder.ingest(rawCalls)
@@ -529,7 +402,7 @@ func (s *chatStreamer) executeTool(ctx context.Context, registry *sessionRegistr
 // not opt into the marker stream, this is a no-op so strict SDKs see a
 // pristine spec-conformant stream.
 func (s *chatStreamer) emitTinfoilEventMarker(id, status string, action map[string]any, reason string, sources []toolCallSource) {
-	if !s.eventsEnabled {
+	if !s.eventFlags.webSearch {
 		return
 	}
 	marker := tinfoilEventMarker(id, status, action, reason, sources)
@@ -574,6 +447,424 @@ func (s *chatStreamer) emitData(body map[string]any) {
 	s.flusher.Flush()
 }
 
+// finalize emits any remaining buffered citations, forwards client-owned
+// tool_calls the model asked the user to execute, writes the finish-reason
+// chunk and final usage chunk, fires the billing event, and closes the
+// stream with `[DONE]`.
+func (s *chatStreamer) finalize(
+	r *http.Request,
+	em *manager.EnclaveManager,
+	modelName string,
+	result chatIterationResult,
+	clientToolCalls []toolCall,
+) error {
+	// Defensive: in the degenerate case where upstream never emitted a
+	// single chunk before signaling done (and the pump therefore never
+	// ran ensureRoleEmitted), the finish chunk below would be the
+	// client's first visible delta; still emit role first to keep the
+	// protocol contract.
+	s.ensureRoleEmitted()
+	s.flushCitations()
+
+	if len(clientToolCalls) > 0 {
+		raw := rawToolCallsFromParsed(clientToolCalls, result.rawToolCalls)
+		s.writeChunk(map[string]any{
+			"choices": []any{
+				map[string]any{
+					"index": 0,
+					"delta": map[string]any{"tool_calls": raw},
+				},
+			},
+		})
+	}
+
+	finishReason := result.finishReason
+	if len(clientToolCalls) > 0 {
+		finishReason = "tool_calls"
+	} else if finishReason == "" {
+		finishReason = "stop"
+	}
+	s.writeChunk(map[string]any{
+		"choices": []any{
+			map[string]any{
+				"index":         0,
+				"delta":         map[string]any{},
+				"finish_reason": finishReason,
+			},
+		},
+	})
+
+	finalUsage := s.finalUsage()
+	if s.clientRequestedUsage && finalUsage != nil {
+		s.writeChunk(map[string]any{
+			"choices": []any{},
+			"usage":   finalUsage,
+		})
+	}
+
+	if s.writeErr == nil {
+		if _, err := io.WriteString(s.w, "data: [DONE]\n\n"); err != nil {
+			s.writeErr = err
+		} else {
+			s.flusher.Flush()
+		}
+	}
+
+	if s.usageMetricsRequested && finalUsage != nil {
+		s.w.Header().Set(manager.UsageMetricsResponseHeader, formatUsageHeader(usageFromRaw(finalUsage)))
+	}
+	s.emitBillingEvent(r, em, modelName, finalUsage)
+	return s.writeErr
+}
+
+// finalUsage assembles the aggregated usage block across every upstream
+// iteration in the OpenAI Chat Completions shape. The underlying
+// accumulator already handles turns where upstream omitted the usage block
+// entirely by falling back to zero.
+func (s *chatStreamer) finalUsage() map[string]any {
+	usage := s.usageTotals.Usage()
+	if usage == nil {
+		return nil
+	}
+	return map[string]any{
+		"prompt_tokens":     usage.PromptTokens,
+		"completion_tokens": usage.CompletionTokens,
+		"total_tokens":      usage.TotalTokens,
+	}
+}
+
+// terminateWithError is the single point at which a mid-stream upstream
+// failure is surfaced to the client. If no SSE has started yet we return
+// the upstream error to the caller so Handle turns it into a plain JSON
+// error response; otherwise we emit a top-level error data frame in the
+// shape OpenAI SDKs recognize mid-stream, followed by `[DONE]` so SDKs can
+// close their streams cleanly.
+func (s *chatStreamer) terminateWithError(r *http.Request, em *manager.EnclaveManager, modelName string, err error) error {
+	if !s.headersWritten {
+		return err
+	}
+	s.flushCitations()
+	s.emitData(map[string]any{"error": upstreamErrorPayload(err)})
+	if s.writeErr == nil {
+		if _, werr := io.WriteString(s.w, "data: [DONE]\n\n"); werr != nil {
+			s.writeErr = werr
+		} else {
+			s.flusher.Flush()
+		}
+	}
+	s.emitBillingEvent(r, em, modelName, s.finalUsage())
+	return nil
+}
+
+// ---------------------------------------------------------------------------
+// chatToolCallBuilder methods
+// ---------------------------------------------------------------------------
+
+// ingest merges one tool_call delta list into the builder's running state.
+// It preserves the original wire-format fragment in s.raw so the assistant
+// message emitted on the next iteration carries bytewise-identical tool
+// calls to whatever upstream produced.
+func (b *chatToolCallBuilder) ingest(rawCalls []any) {
+	for _, item := range rawCalls {
+		m, _ := item.(map[string]any)
+		if m == nil {
+			continue
+		}
+		idx := int(numberValue(m["index"]))
+		if idx < 0 {
+			// OpenAI's streaming tool_call protocol uses a 0-based
+			// index; a negative value is either a protocol bug or
+			// a hostile/malformed upstream frame. Dropping the
+			// delta is safer than indexing into b.entries with a
+			// negative int, which would panic the streamer.
+			continue
+		}
+		for len(b.entries) <= idx {
+			b.entries = append(b.entries, &chatToolCallEntry{index: len(b.entries)})
+		}
+		entry := b.entries[idx]
+		if id := stringValue(m["id"]); id != "" {
+			entry.id = id
+		}
+		if t := stringValue(m["type"]); t != "" {
+			entry.toolType = t
+		}
+		if fn, ok := m["function"].(map[string]any); ok {
+			if name := stringValue(fn["name"]); name != "" {
+				entry.functionName = name
+			}
+			if args, ok := fn["arguments"].(string); ok && args != "" {
+				entry.arguments = append(entry.arguments, args...)
+			}
+		}
+	}
+}
+
+// toolCalls returns the assembled tool_call objects as the parsed router
+// type. Arguments are JSON-decoded when possible; malformed arguments are
+// preserved as an empty map so the loop can still surface the tool call
+// rather than silently drop it.
+func (b *chatToolCallBuilder) toolCalls() []toolCall {
+	out := make([]toolCall, 0, len(b.entries))
+	for _, entry := range b.entries {
+		if entry == nil || entry.functionName == "" {
+			continue
+		}
+		args := map[string]any{}
+		if len(entry.arguments) > 0 {
+			raw, _ := sanitizeToolCallArgumentsJSON(entry.arguments)
+			_ = json.Unmarshal(raw, &args)
+		}
+		out = append(out, toolCall{
+			id:        entry.id,
+			name:      entry.functionName,
+			arguments: args,
+		})
+	}
+	return out
+}
+
+// raw returns the tool_calls array in the wire shape OpenAI uses for
+// non-streaming completions. Used to rehydrate the assistant message we
+// prepend to the next iteration's messages array.
+func (b *chatToolCallBuilder) raw() []any {
+	calls := make([]any, 0, len(b.entries))
+	for _, entry := range b.entries {
+		if entry == nil || entry.functionName == "" {
+			continue
+		}
+		toolType := entry.toolType
+		if toolType == "" {
+			toolType = "function"
+		}
+		calls = append(calls, map[string]any{
+			"id":   entry.id,
+			"type": toolType,
+			"function": map[string]any{
+				"name":      entry.functionName,
+				"arguments": string(entry.arguments),
+			},
+		})
+	}
+	return calls
+}
+
+// ---------------------------------------------------------------------------
+// Free functions
+// ---------------------------------------------------------------------------
+
+// runChatStreaming is the streaming entry point for the Chat Completions
+// route. It drives the same tool loop as runChatLoop but consumes the
+// upstream model's stream incrementally so that assistant content deltas
+// and citation annotations are forwarded to the client live rather than
+// synthesized at the end of the turn. When the client opted into the
+// router-owned `X-Tinfoil-Events: web_search` marker stream, router-owned
+// web_search tool-call progress is surfaced inline via
+// `<tinfoil-event>...</tinfoil-event>` markers carried inside the normal
+// `delta.content` text. Clients that did not opt in see a pristine
+// spec-conformant chat.completion.chunk stream with no markers.
+//
+// The function assumes isStream(body) is already true. Non-streaming callers
+// continue to go through runChatLoop so existing behavior is preserved for
+// SDKs that do not want SSE back.
+func runChatStreaming(
+	ctx context.Context,
+	w http.ResponseWriter,
+	r *http.Request,
+	em *manager.EnclaveManager,
+	registry *sessionRegistry,
+	body map[string]any,
+	modelName string,
+	requestHeaders http.Header,
+	prompt *mcp.GetPromptResult,
+	dl *devLog,
+) error {
+	flusher, ok := w.(http.Flusher)
+	if !ok {
+		return fmt.Errorf("streaming not supported")
+	}
+
+	tid := debugTraceID()
+	searchOpts := parseChatWebSearchOptions(body)
+	tools := registry.allTools()
+	ownedTools := registry.ownedTools()
+	toolSchemas := schemaLookup(tools)
+	reqBody := cloneJSONMap(body)
+	delete(reqBody, "web_search_options")
+	delete(reqBody, "filters")
+	delete(reqBody, "pii_check_options")
+	delete(reqBody, "prompt_injection_check_options")
+	stripRouterOwnedIncludes(reqBody)
+	reqBody["stream"] = true
+	reqBody["stream_options"] = map[string]any{"include_usage": true}
+	applyParallelToolCallsPolicy(reqBody)
+	reqBody["tools"] = append(existingTools(reqBody["tools"]), chatTools(tools)...)
+	reqBody["messages"] = prependChatPrompt(prompt, reqBody["messages"])
+
+	usageMetricsRequested := r.Header.Get(manager.UsageMetricsRequestHeader) == "true"
+	clientRequestedUsage := r.Header.Get("X-Tinfoil-Client-Requested-Usage") == "true"
+	eventFlags := parseTinfoilEventFlags(r.Header)
+
+	if debugEnabled {
+		ownedNames := make([]string, 0, len(ownedTools))
+		for n := range ownedTools {
+			ownedNames = append(ownedNames, n)
+		}
+		debugLogf("toolruntime:%s chatstream.start model=%s search_ctx=%q profiles=%v ownedTools=%v schemas=%d",
+			tid, modelName, searchOpts.searchContextSize, registry.profileNames(), ownedNames, len(toolSchemas))
+	}
+
+	streamer := &chatStreamer{
+		streamBase: streamBase{
+			w:                     w,
+			flusher:               flusher,
+			usageMetricsRequested: usageMetricsRequested,
+			citations:             &citationState{nextIndex: 1},
+			toolCalls:             &toolCallLog{},
+			usageTotals:           &usageAccumulator{},
+		},
+		clientRequestedUsage: clientRequestedUsage,
+		eventFlags:           eventFlags,
+		emitter:              nil,
+	}
+	streamer.emitter = newCitationEmitter(streamer.citations)
+
+	for i := 0; i < maxToolIterations; i++ {
+		if msgs, ok := reqBody["messages"].([]any); ok {
+			if n := sanitizeAssistantToolCallsInMessages(msgs); n > 0 {
+				debugLogf("toolruntime:%s chatstream.iter=%d history_sanitized count=%d", tid, i, n)
+			}
+		}
+		if debugEnabled {
+			msgs, _ := reqBody["messages"].([]any)
+			debugLogf("toolruntime:%s chatstream.iter=%d upstream.request messages_n=%d history=%s", tid, i, len(msgs), debugMessagesSummary(msgs, 200))
+		}
+		iterStart := time.Now()
+		result, err := streamer.runIteration(ctx, em, modelName, reqBody, requestHeaders)
+		if err != nil {
+			debugLogf("toolruntime:%s chatstream.iter=%d upstream.error elapsed=%s err=%v", tid, i, time.Since(iterStart), err)
+			return streamer.terminateWithError(r, em, modelName, err)
+		}
+		// If the client has disconnected, stop before spending another
+		// round of MCP tool calls and upstream tokens on work nobody is
+		// listening for.
+		if streamer.writeErr != nil {
+			debugLogf("toolruntime:%s chatstream.client_disconnected at iter=%d", tid, i)
+			streamer.emitBillingEvent(r, em, modelName, streamer.finalUsage())
+			return nil
+		}
+
+		// Dev log: turn header + tokens + thinking + content
+		dl.WriteTurnHeader(i + 1)
+		dl.WriteTokens(result.usage)
+		dl.WriteThinking(result.reasoning)
+		dl.WriteContent(result.content)
+
+		// Mid-turn client tool_calls are intentionally dropped; only the
+		// final turn (no router tool calls to resolve) forwards them to
+		// the client. This preserves parity with runChatLoop.
+		routerToolCalls, clientToolCalls := splitToolCalls(ownedTools, result.toolCalls)
+		if debugEnabled {
+			toolNames := make([]string, 0, len(result.toolCalls))
+			for _, c := range result.toolCalls {
+				toolNames = append(toolNames, c.name)
+			}
+			debugLogf("toolruntime:%s chatstream.iter=%d upstream.response elapsed=%s finish=%q total_tool_calls=%d router_tool_calls=%d client_tool_calls=%d tool_names=%v content=%q",
+				tid, i, time.Since(iterStart), result.finishReason, len(result.toolCalls), len(routerToolCalls), len(clientToolCalls), toolNames, debugPreview(result.content, 240))
+		}
+		if len(routerToolCalls) == 0 {
+			debugLogf("toolruntime:%s chatstream.done iterations=%d (no router tool calls)", tid, i+1)
+			dl.WriteFinish(result.finishReason)
+			return streamer.finalize(r, em, modelName, result, clientToolCalls)
+		}
+
+		mixedTurn := len(clientToolCalls) > 0
+		dl.WriteToolCalls(routerToolCalls)
+		tracePhase := fmt.Sprintf("chatstream.iter=%d", i)
+		var messages []any
+		if !mixedTurn {
+			messages, _ = reqBody["messages"].([]any)
+			messages = append(messages, result.assistantMessage())
+		}
+		for _, call := range routerToolCalls {
+			tstart := time.Now()
+			output := resolveStreamingRouterToolCall(
+				ctx, call, searchOpts, toolSchemas, streamer.citations, streamer.toolCalls,
+				func(ctx context.Context, call toolCall) (string, error) {
+					return streamer.executeTool(ctx, registry, call)
+				},
+				tracePhase, tid,
+			)
+			dl.WriteToolExec(call.name, call.arguments, output, time.Since(tstart), "")
+			if !mixedTurn {
+				messages = append(messages, map[string]any{
+					"role":         "tool",
+					"tool_call_id": call.id,
+					"content":      output,
+				})
+			}
+		}
+		// Mixed turn: see the mixed-turn contract on runToolLoop.
+		if mixedTurn {
+			debugLogf("toolruntime:%s chatstream.mixed_turn iterations=%d (router+client calls in one turn; finalizing)", tid, i+1)
+			return streamer.finalize(r, em, modelName, result, clientToolCalls)
+		}
+		reqBody["messages"] = messages
+	}
+	debugLogf("toolruntime:%s chatstream.exhausted iterations=%d forcing final answer", tid, maxToolIterations)
+
+	// Exhausted the tool budget; force a final-answer turn with tools removed
+	// and stream it to the client directly.
+	finalBody := forcedFinalChatRequest(reqBody)
+	finalBody["stream"] = true
+	finalBody["stream_options"] = map[string]any{"include_usage": true}
+	result, err := streamer.runIteration(ctx, em, modelName, finalBody, requestHeaders)
+	if err != nil {
+		return streamer.terminateWithError(r, em, modelName, err)
+	}
+	return streamer.finalize(r, em, modelName, result, nil)
+}
+
+// rawToolCallsFromParsed filters the streamed raw tool_calls down to the
+// client-owned ones and re-indexes them starting at zero so the final
+// stream chunk emits a delta.tool_calls array in the shape clients expect:
+// contiguous `index` values identifying positions in the client-visible
+// tool_calls list. The `id`, `type`, and `function` fields are preserved
+// byte-for-byte from what upstream produced.
+func rawToolCallsFromParsed(parsed []toolCall, raw []any) []any {
+	filtered := make([]any, 0, len(parsed))
+	next := 0
+	for _, item := range raw {
+		source, _ := item.(map[string]any)
+		if source == nil || next >= len(parsed) || !rawToolCallMatchesParsed(source, parsed[next]) {
+			continue
+		}
+		reindexed := map[string]any{
+			"index":    len(filtered),
+			"id":       source["id"],
+			"type":     source["type"],
+			"function": source["function"],
+		}
+		filtered = append(filtered, reindexed)
+		next++
+	}
+	return filtered
+}
+
+func rawToolCallMatchesParsed(source map[string]any, call toolCall) bool {
+	if source == nil {
+		return false
+	}
+	function, _ := source["function"].(map[string]any)
+	if function == nil || stringValue(function["name"]) != call.name {
+		return false
+	}
+	if call.id != "" && stringValue(source["id"]) != call.id {
+		return false
+	}
+	return true
+}
+
 // nestedAnnotationsFromMatches converts the internal citationAnnotation
 // slice into the nested url_citation shape documented by OpenAI for Chat
 // Completions responses so SDKs and UIs that already consume the final
@@ -597,6 +888,26 @@ func nestedAnnotationsFromMatches(matches []citationAnnotation) []any {
 	return out
 }
 
+// upstreamErrorPayload extracts the most informative structured error
+// object we can show the client. If the underlying error is an
+// *upstreamError whose body is JSON with an `error` field, we return that
+// object verbatim (preserving the upstream's own `code`, `type`, `param`,
+// etc.). Otherwise we fall back to a minimal `{message, type}` envelope.
+func upstreamErrorPayload(err error) map[string]any {
+	if upErr, ok := err.(*upstreamError); ok && len(upErr.body) > 0 {
+		var parsed map[string]any
+		if json.Unmarshal(upErr.body, &parsed) == nil {
+			if inner, ok := parsed["error"].(map[string]any); ok {
+				return inner
+			}
+		}
+	}
+	return map[string]any{
+		"message": err.Error(),
+		"type":    "upstream_error",
+	}
+}
+
 // mustMarshal serializes the given value to JSON, returning an empty byte
 // slice on failure. Used by error-surfacing helpers where failure to
 // marshal is already a lost cause; returning empty body preserves the
@@ -608,5 +919,3 @@ func mustMarshal(value any) []byte {
 	}
 	return data
 }
-
-
