+++ toolruntime/chat_stream_test.go
@@ -25,6 +25,7 @@ func newTestChatStreamer(t *testing.T) (*chatStreamer, *httptest.ResponseRecorde
 			flusher:               rec,
 			usageMetricsRequested: true,
 			citations:             &citationState{nextIndex: 1},
+			toolCalls:             &toolCallLog{},
 			usageTotals:           &usageAccumulator{},
 			model:                 "gpt-oss-120b",
 			headersWritten:        true,
@@ -439,6 +440,7 @@ func TestChatStreamerPumpAbortsOnClientDisconnect(t *testing.T) {
 			w:              w,
 			flusher:        w,
 			citations:      &citationState{nextIndex: 1},
+			toolCalls:      &toolCallLog{},
 			usageTotals:    &usageAccumulator{},
 			headersWritten: true,
 			model:          "gpt-oss-120b",
