+++ toolruntime/events.go
@@ -25,81 +25,42 @@ const (
 	tinfoilEventPayloadType = "tinfoil.web_search_call"
 )
 
-// webSearchCallEvent builds a single web_search_call output item for the
-// non-streaming `output[]` array. The `blocked` status is collapsed to
-// the spec-valid `failed` because OpenAI's web_search_call.status enum
-// has no `blocked` slot; the unfiltered status still rides on the
-// `_tinfoil` sidecar for clients that want to distinguish a safety
-// block from a generic failure.
-func webSearchCallEvent(id, status, errorCode string, action map[string]any) map[string]any {
-	item := map[string]any{
-		"type":   "web_search_call",
-		"id":     id,
-		"status": status,
-		"action": action,
-	}
-	if status == "blocked" {
-		item["status"] = "failed"
-	}
-	if sidecar := tinfoilSidecar(status, errorCode); sidecar != nil {
-		item["_tinfoil"] = sidecar
-	}
-	return item
+// tinfoilEventFlags tracks which opt-in marker families the caller
+// requested via the `X-Tinfoil-Events` header. Each field gates the
+// corresponding marker type so a client that only understands web
+// search markers does not see unrelated markers.
+type tinfoilEventFlags struct {
+	webSearch bool
 }
 
-// tinfoilSidecar builds the `_tinfoil` vendor-extension field that rides
-// alongside a web_search_call item on the Responses API. Strict OpenAI
-// SDKs ignore unknown object fields, so this is invisible to clients
-// that don't opt into reading it; clients that do (tinfoil-webapp,
-// tinfoil-ios, anyone building a richer error UI on top of Tinfoil) can
-// consume it directly off `item._tinfoil` with no additional opt-in.
-//
-// The sidecar carries ONLY information the spec cannot express:
-//   - `status`: the unfiltered router status, which may be `blocked`
-//     (distinct from the spec-valid `failed` that rides on the envelope
-//     `status` field). Present only when the router status differs from
-//     the envelope status, i.e., only on `blocked` today.
-//   - `error.code`: an opaque router error code (e.g.
-//     `blocked_by_safety_filter`) for clients that want to branch UI on
-//     the specific reason. Present only when the tool call errored.
-//
-// Returns nil when there is nothing worth surfacing (the default, happy
-// path) so the `_tinfoil` field is simply omitted from the item and the
-// serialized JSON stays minimal for successful searches.
-func tinfoilSidecar(rawStatus, errorCode string) map[string]any {
-	if rawStatus != "blocked" && errorCode == "" {
-		return nil
-	}
-	sidecar := map[string]any{}
-	if rawStatus == "blocked" {
-		sidecar["status"] = rawStatus
-	}
-	if errorCode != "" {
-		sidecar["error"] = map[string]any{"code": errorCode}
-	}
-	return sidecar
-}
-
-// tinfoilEventsEnabled reports whether the caller opted into the
-// router-owned tinfoil-event marker stream for web-search tool calls.
-// The header value is a comma-separated list of families the client is
-// prepared to parse. We match case-insensitively and tolerate whitespace.
-// A missing or empty header means the client gets a pristine
+// parseTinfoilEventFlags parses the `X-Tinfoil-Events` header into per-family
+// booleans. The header value is a comma-separated list of families the
+// client is prepared to parse. We match case-insensitively and tolerate
+// whitespace. A missing or empty header means the client gets a pristine
 // spec-conformant stream with no markers at all.
-func tinfoilEventsEnabled(h http.Header) bool {
+func parseTinfoilEventFlags(h http.Header) tinfoilEventFlags {
 	if h == nil {
-		return false
+		return tinfoilEventFlags{}
 	}
 	raw := h.Get(tinfoilEventsHeader)
 	if raw == "" {
-		return false
+		return tinfoilEventFlags{}
 	}
+	var flags tinfoilEventFlags
 	for _, part := range strings.Split(raw, ",") {
-		if strings.EqualFold(strings.TrimSpace(part), tinfoilEventsWebSearch) {
-			return true
+		family := strings.TrimSpace(part)
+		if strings.EqualFold(family, tinfoilEventsWebSearch) {
+			flags.webSearch = true
 		}
 	}
-	return false
+	return flags
+}
+
+// tinfoilEventsEnabled reports whether the caller opted into web-search
+// tinfoil-event markers. Retained for call sites that only need the
+// web-search gate.
+func tinfoilEventsEnabled(h http.Header) bool {
+	return parseTinfoilEventFlags(h).webSearch
 }
 
 // tinfoilEventMarker renders a single progress event as a text marker:
@@ -205,17 +166,4 @@ func tinfoilEventMarkersForRecords(records []toolCallRecord) string {
 	return builder.String()
 }
 
-// fetchArgumentURLs extracts the string URLs the model handed the fetch tool.
-func fetchArgumentURLs(arguments map[string]any) []string {
-	raw, ok := arguments["urls"].([]any)
-	if !ok {
-		return nil
-	}
-	urls := make([]string, 0, len(raw))
-	for _, item := range raw {
-		if url := strings.TrimSpace(stringValue(item)); url != "" {
-			urls = append(urls, url)
-		}
-	}
-	return urls
-}
+
