+++ toolruntime/responses_stream_test.go
@@ -29,6 +29,7 @@ func newTestResponsesStreamer(t *testing.T) (*responsesStreamer, *httptest.Respo
 			flusher:               rec,
 			usageMetricsRequested: true,
 			citations:             &citationState{nextIndex: 1},
+			toolCalls:             &toolCallLog{},
 			usageTotals:           &usageAccumulator{},
 			model:                 "gpt-oss-120b",
 			headersWritten:        true,
@@ -605,6 +606,7 @@ func TestResponsesStreamerPumpAbortsOnClientDisconnect(t *testing.T) {
 			w:              w,
 			flusher:        w,
 			citations:      &citationState{nextIndex: 1},
+			toolCalls:      &toolCallLog{},
 			usageTotals:    &usageAccumulator{},
 			model:          "gpt-oss-120b",
 			headersWritten: true,
